requirements.txt
mlflow>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.1.0
requests>=2.28.0

# Data drift / monitoring
evidently>=0.3.0

# DagsHub client for experiment tracking/metadata
dagshub>=0.6.3

# environment-specific
# dotenv is used by fetch_waqi_station_details_sea.py to load WAQI_TOKEN
python-dotenv>=0.21.0